+++ Scripts/Global.gd
@@ -16,7 +16,7 @@ var canvases := []
 # warning-ignore:unused_class_variable
 var hidden_canvases := []
 enum PRESSURE_SENSITIVITY {NONE, ALPHA, SIZE}
-var pressure_sensitivity_mode = PRESSURE_SENSITIVITY.NONE
+var pressure_sensitivity_mode = PRESSURE_SENSITIVITY.ALPHA
 var smooth_zoom := true
 var left_cursor_tool_texture : ImageTexture
 var right_cursor_tool_texture : ImageTexture